+++ desktop/preload.js
@@ -3,5 +3,6 @@ const { contextBridge, ipcRenderer } = require('electron');
 contextBridge.exposeInMainWorld('sneupDesktop', {
   platform: process.platform,
   saveStartupMode: (startupMode) => ipcRenderer.invoke('sneup:save-startup-mode', startupMode),
-  restart: () => ipcRenderer.invoke('sneup:restart')
+  restart: () => ipcRenderer.invoke('sneup:restart'),
+  createSupportBundle: () => ipcRenderer.invoke('sneup:create-support-bundle')
 });